+++ src/index.ts
@@ -29,6 +29,21 @@ export {
   DividerPropsWithDefaults,
 } from './components/Divider/Divider'
 
+export {
+  default as Dropdown,
+  DropdownProps,
+  A11yStatusMessageOptions,
+} from './components/Dropdown/Dropdown'
+
+export { default as DropdownItem, DropdownItemProps } from './components/Dropdown/DropdownItem'
+
+export { default as DropdownLabel, DropdownLabelProps } from './components/Dropdown/DropdownLabel'
+
+export {
+  default as DropdownSearchInput,
+  DropdownSearchInputProps,
+} from './components/Dropdown/DropdownSearchInput'
+
 export { default as Form, FormProps } from './components/Form/Form'
 export { default as FormField, FormFieldProps } from './components/Form/FormField'