+++ src/lib/accessibility/index.ts
@@ -1,7 +1,6 @@
 export { default as defaultBehavior } from './Behaviors/defaultBehavior'
 export { default as attachmentBehavior } from './Behaviors/Attachment/attachmentBehavior'
 export { default as buttonBehavior } from './Behaviors/Button/buttonBehavior'
-export { default as buttonGroupBehavior } from './Behaviors/Button/buttonGroupBehavior'
 export { default as toggleButtonBehavior } from './Behaviors/Button/toggleButtonBehavior'
 export { default as imageBehavior } from './Behaviors/Image/imageBehavior'
 export { default as menuBehavior } from './Behaviors/Menu/menuBehavior'