+++ src/lib/index.ts
@@ -3,7 +3,6 @@ import * as commonPropTypes from './commonPropTypes'
 
 export { default as AutoControlledComponent } from './AutoControlledComponent'
 export { default as childrenExist } from './childrenExist'
-export * from './colorUtils'
 export { default as UIComponent } from './UIComponent'
 export { EventStack } from './eventStack'
 export { felaRenderer, felaRtlRenderer } from './felaRenderer'
@@ -28,7 +27,6 @@ export {
 } from './htmlPropsUtils'
 
 export { default as isBrowser } from './isBrowser'
-export { default as typescriptUtils } from './typescriptUtils'
 export { default as doesNodeContainClick } from './doesNodeContainClick'
 export { default as leven } from './leven'
 
@@ -36,6 +34,7 @@ export { pxToRem } from './fontSizeUtility'
 export { customPropTypes }
 export { default as createAnimationStyles } from './createAnimationStyles'
 export { default as createComponent } from './createStardustComponent'
+export * from './whatInput'
 
 export * from './commonPropInterfaces'
 export { commonPropTypes }