+++ src/themes/teams/components/Icon/svg/ProcessedIcons/icons-msft-canvas.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { TeamsSvgIconSpec } from '../types'
+import { TeamsProcessedSvgIconSpec } from '../types'
 
 export default {
   icon: ({ classes }) => (
@@ -21,4 +21,4 @@ export default {
     </svg>
   ),
   styles: {},
-} as TeamsSvgIconSpec
+} as TeamsProcessedSvgIconSpec