+++ test/specs/behaviors/behavior-test.tsx
@@ -9,7 +9,6 @@ import {
   basicListBehavior,
   basicListItemBehavior,
   buttonBehavior,
-  buttonGroupBehavior,
   iconBehavior,
   imageBehavior,
   inputBehavior,
@@ -42,7 +41,6 @@ testHelper.addBehavior('attachmentBehavior', attachmentBehavior)
 testHelper.addBehavior('basicListBehavior', basicListBehavior)
 testHelper.addBehavior('basicListItemBehavior', basicListItemBehavior)
 testHelper.addBehavior('buttonBehavior', buttonBehavior)
-testHelper.addBehavior('buttonGroupBehavior', buttonGroupBehavior)
 testHelper.addBehavior('iconBehavior', iconBehavior)
 testHelper.addBehavior('inputBehavior', inputBehavior)
 testHelper.addBehavior('imageBehavior', imageBehavior)