+++ test/specs/commonTests/implementsWrapperProp.tsx
@@ -7,14 +7,14 @@ import { ShorthandValue } from 'utils'
 
 export interface ImplementsWrapperPropOptions {
   wrapppedComponentSelector: any
-  wrappperComponentSelector?: any
+  WrapperComponent?: any
 }
 
 const implementsWrapperProp = <P extends { wrapper: ShorthandValue }>(
   Component: React.ReactType<P>,
   options: ImplementsWrapperPropOptions,
 ) => {
-  const { wrapppedComponentSelector, wrappperComponentSelector = Slot.defaultProps.as } = options
+  const { wrapppedComponentSelector, WrapperComponent = Slot } = options
 
   const wrapperTests = (wrapper: ReactWrapper) => {
     expect(wrapper.length).toBeGreaterThan(0)
@@ -23,11 +23,7 @@ const implementsWrapperProp = <P extends { wrapper: ShorthandValue }>(
 
   describe('"wrapper" prop', () => {
     it('wraps the component by default', () => {
-      wrapperTests(mount(<Component />).find(wrappperComponentSelector))
-    })
-
-    it('wraps the component with a custom element', () => {
-      wrapperTests(mount(<Component wrapper={<span />} />).find('span'))
+      wrapperTests(mount(<Component />).find(WrapperComponent))
     })
 
     it('wraps the component with a custom element using "as" prop', () => {